requirements.txt
albumentations==1.3.0
dlib==19.24.0
facenet_pytorch==2.5.2
kornia==0.6.8
lpips==0.1.4
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.3.0
scikit_image==0.19.2
skimage==0.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
